+++ Centipede/Centipede/GameObject.h
@@ -13,14 +13,14 @@ public:
 	~GameObject();
 	int getPointValue();
 	unsigned int getHealth() const;
-	sf::Vector2i getVelocity() const;
-	sf::Vector2i getPosition() const;
+	sf::Vector2f getVelocity() const;
+	sf::Vector2f getPosition() const;
 	void setTexture(const char *);
 	virtual void update(CentipedeGame*) = 0;
 	void render(sf::RenderTexture &);
 	virtual void collideWith(GameObject *);
 	//virtual void collideWith(std::vector<GameObject *> *);
-	sf::Vector2i getNearestCellPos(sf::Vector2i);
+	sf::Vector2f getNearestCellPos(sf::Vector2f);
 
 	static sf::Vector2u oWD;
 	static sf::Vector2i interval;
@@ -31,15 +31,15 @@ public:
 
 protected:
 	void setPointValue(unsigned int);
-	void setVelocity(sf::Vector2i);
+	void setVelocity(sf::Vector2f);
 	void addHealth(int);
 	void setPixels();
 
 	sf::Sprite object;
 	unsigned int health = 1;
 	unsigned int pointValue = 0;
-	sf::Vector2i velocity;
-	sf::Vector2i currentPosition;
+	sf::Vector2f velocity;
+	sf::Vector2f currentPosition;
 	sf::Texture textureTemp;
 
 	sf::Sound soundPlayer;
